move profile page to fragment
