+++ .idea/misc.xml
@@ -14,6 +14,7 @@
         <entry key="app/src/main/res/layout/activity_test.xml" value="0.40260416666666665" />
         <entry key="app/src/main/res/layout/balance_activity.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/layout/custom_marker_view.xml" value="0.40260416666666665" />
+        <entry key="app/src/main/res/layout/fragment_profile.xml" value="0.3304831124412142" />
         <entry key="app/src/main/res/layout/historical_balance_graph.xml" value="0.4036458333333333" />
         <entry key="app/src/main/res/layout/item_transaction.xml" value="0.45" />
         <entry key="app/src/main/res/menu/menu_bottom_navigation.xml" value="0.3304831124412142" />
