+++ app/src/main/res/menu/menu_bottom_navigation.xml
@@ -2,21 +2,15 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
-        android:id="@+id/action_favorites"
+        android:id="@+id/action_profile"
         android:enabled="true"
         android:icon="@drawable/common_full_open_on_phone"
-        android:title="queso"
+        android:title="Profile"
         app:showAsAction="ifRoom" />
     <item
-        android:id="@+id/action_schedules"
+        android:id="@+id/action_search"
         android:enabled="true"
         android:icon="@drawable/ic_baseline_close_24"
-        android:title="butter"
-        app:showAsAction="ifRoom" />
-    <item
-        android:id="@+id/action_music"
-        android:enabled="true"
-        android:icon="@drawable/common_google_signin_btn_icon_dark"
-        android:title="cheese"
+        android:title="Search"
         app:showAsAction="ifRoom" />
 </menu>
